compile check

+++ be/src/util/runtime_profile.h
@@ -47,6 +47,7 @@
 #include "util/stopwatch.hpp"
 
 namespace doris {
+#include "common/compile_check_begin.h"
 class TRuntimeProfileNode;
 class TRuntimeProfileTree;
 class RuntimeProfileCounterTreeNode;
@@ -606,7 +607,7 @@ public:
     void get_all_children(std::vector<RuntimeProfile*>* children);
 
     // Returns the number of counters in this profile
-    int num_counters() const { return _counter_map.size(); }
+    size_t num_counters() const { return _counter_map.size(); }
 
     // Returns name of this profile
     const std::string& name() const { return _name; }
@@ -822,5 +823,5 @@ private:
     T _sw;
     C* _counter = nullptr;
 };
-
+#include "common/compile_check_end.h"
 } // namespace doris
